rename createDataServiceMiddleware to createDataService
