+++ src/createDataService/index.js
@@ -1,0 +1,2 @@
+export { default } from './createDataService';
+export * from './createDataService';
