--- src/createDataServiceMiddleware/index.js
@@ -1,2 +1,0 @@
-export { default } from './createDataServiceMiddleware';
-export * from './createDataServiceMiddleware';
